SQL queries
For anyone interested in practicing SQL queries.

In this video I will use the MySQL Workbench to solve a query.

The sample consume database will be a college and tables: lecturer, student and department.

You are welcome to access the code solution and the question in github:
https://github.com/racheliver

"Sql collage"
And also follow in linkedin:

https://www.linkedin.com/in/racheli-verechzon-5343bb123/

recommends entering the online examination database of the Hebrew University.